requirements.txt
blinker==1.7.0
certifi==2024.07.04
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.6.1
Flask==3.0.2
idna==3.7
itsdangerous==2.1.2
Jinja2==3.1.4
MarkupSafe==2.1.5
pillow==10.3.0
pymongo==4.6.3
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.0
urllib3==2.2.2
Werkzeug==3.0.3
aiohttp>=3.9.4
bs4==0.0.2
lxml==5.1.0
pillow==10.3.0
flask-wtf==1.2.1
cloudinary==1.39.1
six==1.16.0
